+++ src/main/java/com/lgmpjt/websnsspringboot/application/port/in/BoardCommandUseCase.java
@@ -1,6 +1,6 @@
 package com.lgmpjt.websnsspringboot.application.port.in;
 
-import com.lgmpjt.websnsspringboot.adapter.out.entity.Board;
+import com.lgmpjt.websnsspringboot.adapter.out.persistence.entity.Board;
 import com.lgmpjt.websnsspringboot.application.port.in.dto.BoardCreateDto;
 import com.lgmpjt.websnsspringboot.application.port.in.dto.BoardDto;
 import org.springframework.transaction.annotation.Transactional;
