+++ src/main/java/com/lgmpjt/websnsspringboot/application/port/in/LikeCommandUseCase.java
@@ -1,6 +1,6 @@
 package com.lgmpjt.websnsspringboot.application.port.in;
 
-import com.lgmpjt.websnsspringboot.adapter.out.entity.LikeEntity;
+import com.lgmpjt.websnsspringboot.adapter.out.persistence.entity.LikeEntity;
 import org.springframework.transaction.annotation.Transactional;
 
 public interface LikeCommandUseCase {
