+++ src/main/java/com/lgmpjt/websnsspringboot/application/port/in/UserCommandUseCase.java
@@ -1,6 +1,6 @@
 package com.lgmpjt.websnsspringboot.application.port.in;
 
-import com.lgmpjt.websnsspringboot.adapter.out.entity.User;
+import com.lgmpjt.websnsspringboot.adapter.out.persistence.entity.User;
 import com.lgmpjt.websnsspringboot.application.port.in.dto.UserCreateDto;
 import com.lgmpjt.websnsspringboot.application.port.in.dto.UserDto;
 import org.springframework.transaction.annotation.Transactional;
