+++ src/main/java/com/lgmpjt/websnsspringboot/infra/repository/BoardJpaRepository.java
@@ -1,6 +1,6 @@
 package com.lgmpjt.websnsspringboot.infra.repository;
 
-import com.lgmpjt.websnsspringboot.adapter.out.entity.Board;
+import com.lgmpjt.websnsspringboot.adapter.out.persistence.entity.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
