+++ src/main/java/com/lgmpjt/websnsspringboot/mapper/FollowMapper.java
@@ -1,7 +1,7 @@
 package com.lgmpjt.websnsspringboot.mapper;
 
+import com.lgmpjt.websnsspringboot.adapter.out.persistence.entity.Follow;
 import com.lgmpjt.websnsspringboot.application.port.in.dto.FollowDto;
-import com.lgmpjt.websnsspringboot.adapter.out.entity.Follow;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 import org.mapstruct.factory.Mappers;
